+++ containers/Dockerfile.combined
@@ -37,7 +37,8 @@ COPY requirements.txt .
 RUN pip install -r requirements.txt
 
 WORKDIR /app
-COPY run_script.py .
+COPY main.py .
+COPY utils.py .
 COPY light_server.py .
 RUN mkdir /data
 RUN mkdir /logs
@@ -46,4 +47,3 @@ RUN mkdir /logs
 ENV PYTHONUNBUFFERED=1
 
 CMD ["python", "light_server.py"]
-# CMD ["uvicorn", "light_server:app", "--host", "0.0.0.0", "--port", "5000", "--reload"]
