+++ containers/Dockerfile.ks3
@@ -1,5 +1,5 @@
 # Spike sorter image
-FROM spikeinterface/kilosort3-compiled-base as ksbase
+FROM spikeinterface/kilosort3-compiled-base:0.2.0 as ksbase
 
 # NVIDIA-ready Image
 FROM nvidia/cuda:11.6.2-base-ubuntu20.04
@@ -12,9 +12,8 @@ RUN apt-get update && \
     apt-get clean && \
     rm -rf /var/lib/apt/lists/*
 
-ENV MINICONDA_VERSION 4.8.2
 ENV CONDA_DIR /home/miniconda3
-ENV LATEST_CONDA_SCRIPT "Miniconda3-py38_$MINICONDA_VERSION-Linux-x86_64.sh"
+ENV LATEST_CONDA_SCRIPT "Miniconda3-py39_23.5.2-0-Linux-x86_64.sh"
 
 RUN wget --quiet https://repo.anaconda.com/miniconda/$LATEST_CONDA_SCRIPT -O ~/miniconda.sh && \
     bash ~/miniconda.sh -b -p $CONDA_DIR && \
@@ -29,20 +28,20 @@ ENV PATH="/opt/matlabruntime:${PATH}"
 COPY --from=ksbase /usr/lib/x86_64-linux-gnu/libXt.so.6 /usr/lib/x86_64-linux-gnu/libXt.so.6
 COPY --from=ksbase /usr/lib/x86_64-linux-gnu/libSM.so.6 /usr/lib/x86_64-linux-gnu/libSM.so.6
 COPY --from=ksbase /usr/lib/x86_64-linux-gnu/libICE.so.6 /usr/lib/x86_64-linux-gnu/libICE.so.6
-ENV LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/opt/matlabruntime/v911/runtime/glnxa64:/opt/matlabruntime/v911/bin/glnxa64:/opt/matlabruntime/v911/sys/os/glnxa64:/opt/matlabruntime/v911/sys/opengl/lib/glnxa64:/opt/matlabruntime/v911/extern/bin/glnxa64
-
-# Get Python stdout logs
-ENV PYTHONUNBUFFERED=1
+ENV LD_LIBRARY_PATH=$LD_LIBRARY_PATH:/opt/matlabruntime/R2022b/runtime/glnxa64:/opt/matlabruntime/R2022b/bin/glnxa64:/opt/matlabruntime/R2022b/sys/os/glnxa64:/opt/matlabruntime/R2022b/sys/opengl/lib/glnxa64:/opt/matlabruntime/R2022b/extern/bin/glnxa64
 
 # Copy requirements and script
 COPY requirements.txt .
 RUN pip install -r requirements.txt
 
 WORKDIR /app
-COPY run_script.py .
+COPY main.py .
+COPY utils.py .
 COPY light_server.py .
 RUN mkdir /data
 RUN mkdir /logs
 
+# Get Python stdout logs
+ENV PYTHONUNBUFFERED=1
+
 CMD ["python", "light_server.py"]
-# ENTRYPOINT ["python", "run_script.py"]
